requirements.txt
FastAPI
sqlalchemy
python_dotenv
asyncpg
psycopg2-binary
uvicorn